continue module shuffling
