+++ src/args.jl
@@ -1680,10 +1680,6 @@ end
 
 # -----------------------------------------------------------------------------
 
-function color_or_nothing!(plotattributes, k::Symbol)
-    plotattributes[k] = (v = plotattributes[k]) === :match ? v : plot_color(v)
-    nothing
-end
 
 # -----------------------------------------------------------------------------
 
@@ -1734,16 +1730,6 @@ const _match_map2 = Dict(
 
 
 # -----------------------------------------------------------------------------
-
-function fg_color(plotattributes::AKW)
-    fg = get(plotattributes, :foreground_color, :auto)
-    if fg === :auto
-        bg = plot_color(get(plotattributes, :background_color, :white))
-        fg = alpha(bg) > 0 && isdark(bg) ? colorant"white" : colorant"black"
-    else
-        plot_color(fg)
-    end
-end
 
 
 # -----------------------------------------------------------------------------
