+++ src/utils.jl
@@ -781,166 +781,6 @@ extend_by_data!(v::AbstractVector, x::AbstractVector) =
 
 # -------------------------------------------------------
 
-function attr!(series::Series; kw...)
-    plotattributes = KW(kw)
-    Plots.preprocess_attributes!(plotattributes)
-    for (k, v) in plotattributes
-        if haskey(_series_defaults, k)
-            series[k] = v
-        else
-            @warn "unused key $k in series attr"
-        end
-    end
-    _series_updated(series[:subplot].plt, series)
-    series
-end
-
-function attr!(sp::Subplot; kw...)
-    plotattributes = KW(kw)
-    Plots.preprocess_attributes!(plotattributes)
-    for (k, v) in plotattributes
-        if haskey(_subplot_defaults, k)
-            sp[k] = v
-        else
-            @warn "unused key $k in subplot attr"
-        end
-    end
-    sp
-end
-
-# -------------------------------------------------------
-# push/append for one series
-
-Base.push!(plt::Plot, args::Real...) = push!(plt, 1, args...)
-Base.push!(plt::Plot, i::Integer, args::Real...) = push!(plt.series_list[i], args...)
-Base.append!(plt::Plot, args::AbstractVector) = append!(plt, 1, args...)
-Base.append!(plt::Plot, i::Integer, args::Real...) = append!(plt.series_list[i], args...)
-
-# tuples
-Base.push!(plt::Plot, t::Tuple) = push!(plt, 1, t...)
-Base.push!(plt::Plot, i::Integer, t::Tuple) = push!(plt, i, t...)
-Base.append!(plt::Plot, t::Tuple) = append!(plt, 1, t...)
-Base.append!(plt::Plot, i::Integer, t::Tuple) = append!(plt, i, t...)
-
-# -------------------------------------------------------
-# push/append for all series
-
-# push y[i] to the ith series
-function Base.push!(plt::Plot, y::AVec)
-    ny = length(y)
-    for i in 1:(plt.n)
-        push!(plt, i, y[mod1(i, ny)])
-    end
-    plt
-end
-
-# push y[i] to the ith series
-# same x for each series
-Base.push!(plt::Plot, x::Real, y::AVec) = push!(plt, [x], y)
-
-# push (x[i], y[i]) to the ith series
-function Base.push!(plt::Plot, x::AVec, y::AVec)
-    nx = length(x)
-    ny = length(y)
-    for i in 1:(plt.n)
-        push!(plt, i, x[mod1(i, nx)], y[mod1(i, ny)])
-    end
-    plt
-end
-
-# push (x[i], y[i], z[i]) to the ith series
-function Base.push!(plt::Plot, x::AVec, y::AVec, z::AVec)
-    nx = length(x)
-    ny = length(y)
-    nz = length(z)
-    for i in 1:(plt.n)
-        push!(plt, i, x[mod1(i, nx)], y[mod1(i, ny)], z[mod1(i, nz)])
-    end
-    plt
-end
-
-# ---------------------------------------------------------------
-
-
-"Smallest x in plot"
-xmin(plt::Plot) = ignorenan_minimum([
-    ignorenan_minimum(series.plotattributes[:x]) for series in plt.series_list
-])
-"Largest x in plot"
-xmax(plt::Plot) = ignorenan_maximum([
-    ignorenan_maximum(series.plotattributes[:x]) for series in plt.series_list
-])
-
-"Extrema of x-values in plot"
-ignorenan_extrema(plt::Plot) = (xmin(plt), xmax(plt))
-
-# ---------------------------------------------------------------
-# get fonts from objects:
-
-plottitlefont(p::Plot) = font(;
-    family = p[:plot_titlefontfamily],
-    pointsize = p[:plot_titlefontsize],
-    valign = p[:plot_titlefontvalign],
-    halign = p[:plot_titlefonthalign],
-    rotation = p[:plot_titlefontrotation],
-    color = p[:plot_titlefontcolor],
-)
-
-colorbartitlefont(sp::Subplot) = font(;
-    family = sp[:colorbar_titlefontfamily],
-    pointsize = sp[:colorbar_titlefontsize],
-    valign = sp[:colorbar_titlefontvalign],
-    halign = sp[:colorbar_titlefonthalign],
-    rotation = sp[:colorbar_titlefontrotation],
-    color = sp[:colorbar_titlefontcolor],
-)
-
-titlefont(sp::Subplot) = font(;
-    family = sp[:titlefontfamily],
-    pointsize = sp[:titlefontsize],
-    valign = sp[:titlefontvalign],
-    halign = sp[:titlefonthalign],
-    rotation = sp[:titlefontrotation],
-    color = sp[:titlefontcolor],
-)
-
-legendfont(sp::Subplot) = font(;
-    family = sp[:legend_font_family],
-    pointsize = sp[:legend_font_pointsize],
-    valign = sp[:legend_font_valign],
-    halign = sp[:legend_font_halign],
-    rotation = sp[:legend_font_rotation],
-    color = sp[:legend_font_color],
-)
-
-legendtitlefont(sp::Subplot) = font(;
-    family = sp[:legend_title_font_family],
-    pointsize = sp[:legend_title_font_pointsize],
-    valign = sp[:legend_title_font_valign],
-    halign = sp[:legend_title_font_halign],
-    rotation = sp[:legend_title_font_rotation],
-    color = sp[:legend_title_font_color],
-)
-
-tickfont(ax::Axis) = font(;
-    family = ax[:tickfontfamily],
-    pointsize = ax[:tickfontsize],
-    valign = ax[:tickfontvalign],
-    halign = ax[:tickfonthalign],
-    rotation = ax[:tickfontrotation],
-    color = ax[:tickfontcolor],
-)
-
-guidefont(ax::Axis) = font(;
-    family = ax[:guidefontfamily],
-    pointsize = ax[:guidefontsize],
-    valign = ax[:guidefontvalign],
-    halign = ax[:guidefonthalign],
-    rotation = ax[:guidefontrotation],
-    color = ax[:guidefontcolor],
-)
-
-# ---------------------------------------------------------------
 # converts unicode scientific notation, as returned by Showoff,
 # to a tex-like format (supported by gr, pyplot, and pgfplots).
